requirements.txt
asgiref==3.8.1
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
